+++ src/editor/core/command/CommandAdapt.ts
@@ -172,6 +172,11 @@ export class CommandAdapt {
     })
   }
 
+  public blur() {
+    this.range.clearRange()
+    this.draw.getCursor().recoveryCursor()
+  }
+
   public undo() {
     const isReadonly = this.draw.isReadonly()
     if (isReadonly) return
@@ -1945,6 +1950,7 @@ export class CommandAdapt {
     // 删除控件
     const control = this.draw.getControl()
     const newIndex = control.removeControl(startIndex)
+    if (newIndex === null) return
     // 重新渲染
     this.range.setRange(newIndex, newIndex)
     this.draw.render({