+++ src/editor/core/contextmenu/menus/globalMenus.ts
@@ -1,9 +1,14 @@
+import { INTERNAL_CONTEXT_MENU_KEY } from '../../../dataset/constant/ContextMenu'
 import { IRegisterContextMenu } from '../../../interface/contextmenu/ContextMenu'
 import { isApple } from '../../../utils/ua'
 import { Command } from '../../command/Command'
+const {
+  GLOBAL: { CUT, COPY, PASTE, SELECT_ALL, PRINT }
+} = INTERNAL_CONTEXT_MENU_KEY
 
 export const globalMenus: IRegisterContextMenu[] = [
   {
+    key: CUT,
     i18nPath: 'contextmenu.global.cut',
     shortCut: `${isApple ? '⌘' : 'Ctrl'} + X`,
     when: payload => {
@@ -14,16 +19,18 @@ export const globalMenus: IRegisterContextMenu[] = [
     }
   },
   {
+    key: COPY,
     i18nPath: 'contextmenu.global.copy',
     shortCut: `${isApple ? '⌘' : 'Ctrl'} + C`,
     when: payload => {
-      return payload.editorHasSelection
+      return payload.editorHasSelection || payload.isCrossRowCol
     },
     callback: (command: Command) => {
       command.executeCopy()
     }
   },
   {
+    key: PASTE,
     i18nPath: 'contextmenu.global.paste',
     shortCut: `${isApple ? '⌘' : 'Ctrl'} + V`,
     when: payload => {
@@ -34,6 +41,7 @@ export const globalMenus: IRegisterContextMenu[] = [
     }
   },
   {
+    key: SELECT_ALL,
     i18nPath: 'contextmenu.global.selectAll',
     shortCut: `${isApple ? '⌘' : 'Ctrl'} + A`,
     when: payload => {
@@ -47,6 +55,7 @@ export const globalMenus: IRegisterContextMenu[] = [
     isDivider: true
   },
   {
+    key: PRINT,
     i18nPath: 'contextmenu.global.print',
     icon: 'print',
     when: () => true,