+++ src/editor/core/contextmenu/menus/hyperlinkMenus.ts
@@ -1,12 +1,17 @@
+import { INTERNAL_CONTEXT_MENU_KEY } from '../../../dataset/constant/ContextMenu'
 import { ElementType } from '../../../dataset/enum/Element'
 import {
   IContextMenuContext,
   IRegisterContextMenu
 } from '../../../interface/contextmenu/ContextMenu'
 import { Command } from '../../command/Command'
+const {
+  HYPERLINK: { DELETE, CANCEL, EDIT }
+} = INTERNAL_CONTEXT_MENU_KEY
 
 export const hyperlinkMenus: IRegisterContextMenu[] = [
   {
+    key: DELETE,
     i18nPath: 'contextmenu.hyperlink.delete',
     when: payload => {
       return (
@@ -19,6 +24,7 @@ export const hyperlinkMenus: IRegisterContextMenu[] = [
     }
   },
   {
+    key: CANCEL,
     i18nPath: 'contextmenu.hyperlink.cancel',
     when: payload => {
       return (
@@ -31,6 +37,7 @@ export const hyperlinkMenus: IRegisterContextMenu[] = [
     }
   },
   {
+    key: EDIT,
     i18nPath: 'contextmenu.hyperlink.edit',
     when: payload => {
       return (